# src/api/gen/forwarder_service/service.go
// Code generated by goa v3.19.1, DO NOT EDIT.
//
// ForwarderService service
//
// Command:
// $ goa gen
// github.com/LeonDavidZipp/StackITCloudAcceleratorCodingChallenge/src/api/design
// --output ./src/api

package forwarderservice

import (
	"context"
)

// The forwarder service forwards warnings to the appropriate channel
type Service interface {
	// Forwards a warning to the appropriate channel
	Forward(context.Context, *Notification) (err error)
}

// APIName is the name of the API as defined in the design.
const APIName = "ForwarderAPI"

// APIVersion is the version of the API as defined in the design.
const APIVersion = "0.0.1"

// ServiceName is the name of the service as defined in the design. This is the
// same value that is set in the endpoint request contexts under the ServiceKey
// key.
const ServiceName = "ForwarderService"

// MethodNames lists the service method names as defined in the design. These
// are the same values that are set in the endpoint request contexts under the
// MethodKey key.
var MethodNames = [1]string{"forward"}

// Invalid notification type
type InvalidNotificationType string

// Notification is the payload type of the ForwarderService service forward
// method.
type Notification struct {
	// The type of notification
	Type string
	// The name of the event
	Name string
	// The description of the event
	Description string
}

// Error returns an error description.
func (e InvalidNotificationType) Error() string {
	return "Invalid notification type"
}

// ErrorName returns "InvalidNotificationType".
//
// Deprecated: Use GoaErrorName - https://github.com/goadesign/goa/issues/3105
func (e InvalidNotificationType) ErrorName() string {
	return e.GoaErrorName()
}

// GoaErrorName returns "InvalidNotificationType".
func (e InvalidNotificationType) GoaErrorName() string {
	return "InvalidNotificationType"
}