# src/api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Forwarder API
    description: The forwarder api forwards warnings to the appropriate channel
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /forward:
        post:
            tags:
                - ForwarderService
            summary: forward ForwarderService
            description: Forwards a warning to the appropriate channel
            operationId: ForwarderService#forward
            parameters:
                - name: ForwardRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/Notification'
                    required:
                        - type
                        - name
                        - description
            responses:
                "201":
                    description: Created response.
                "400":
                    description: Bad Request response.
                    schema:
                        type: string
            schemes:
                - http
definitions:
    Notification:
        title: Notification
        type: object
        properties:
            description:
                type: string
                description: The description of the event
                example: Est minima consequatur corporis.
            name:
                type: string
                description: The name of the event
                example: Deserunt molestiae necessitatibus.
            type:
                type: string
                description: The type of notification
                example: Porro aspernatur ab.
        example:
            description: Est quo.
            name: Ut exercitationem et qui odit.
            type: Officia beatae reprehenderit maiores reiciendis.
        required:
            - type
            - name
            - description